third_party/blink/web_tests/external/wpt/css/selectors/heading-prefixed-expected.txt

This is a testharness.js-based test.
[FAIL] :heading(1) matches <h:h1> in the HTML namespace
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] :heading(2) matches <h:h2> in the HTML namespace
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] :heading(3) matches <h:h3> in the HTML namespace
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] :heading(4) matches <h:h4> in the HTML namespace
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] :heading(5) matches <h:h5> in the HTML namespace
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] :heading(6) matches <h:h6> in the HTML namespace
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
Harness: the test ran to completion.
